implement configuration and solution file parsing
Closes #30
